+++ src/compas_masonry/interactions/__init__.py
@@ -1,0 +1,5 @@
+from .contact import ContactInterface
+
+__all__ = [
+    "ContactInterface",
+]