+++ src/compas_masonry/templates/__init__.py
@@ -1,0 +1,12 @@
+from .template import Template  # noqa: F401
+
+from .arch import ArchTemplate  # noqa: F401
+from .dome import DomeTemplate  # noqa: F401
+from .wall import WallTemplate  # noqa: F401
+
+__all__ = [
+    "Template",
+    "ArchTemplate",
+    "DomeTemplate",
+    "WallTemplate",
+]