+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `compas_masonry.algorithms.mesh_mesh_contacts`.
+* Added `compas_masonry.elements.BlockElement`.
+* Added `compas_masonry.interactions.ContactInterface`.
+* Added `compas_masonry.models.BlockModel`.
+* Added `compas_masonry.templates.Template`.
+* Added `compas_masonry.templates.ArchTemplate`.
+* Added `compas_masonry.viewers.BlockModelViewer`.
+
 ### Changed
 
 ### Removed